+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @glissade/cli
 
+## 0.55.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/lottie@0.55.0-pre.1
+  - @glissade/backend-skia@0.55.0-pre.1
+  - @glissade/core@0.55.0-pre.1
+  - @glissade/interact@0.55.0-pre.1
+  - @glissade/narrate@0.55.0-pre.1
+  - @glissade/player@0.55.0-pre.1
+  - @glissade/scene@0.55.0-pre.1
+  - @glissade/sfx@0.55.0-pre.1
+  - @glissade/svg@0.55.0-pre.1
+
 ## 0.55.0-pre.0
 
 ### Patch Changes